requirements.txt
numpy==1.24.3
pandas==2.1.2
joblib==1.2.0
scikit-learn==1.3.0
Flask==2.2.2
Werkzeug==2.2.2
PyYAML==6.0
requests==2.31.0
groq==0.6.0
gradio==3.50.2
langchain==0.0.320